adopted threadification to new file setup
